// ejercicios/05-heroes-app/src/heroes/types/hero.interface.ts
export interface Hero {
   id: string;
   name: string;
   slug: string;
   alias: string;
   powers: string[];
   description: string;
   strength: number;
   intelligence: number;
   speed: number;
   durability: number;
   team: string;
   image: string;
   firstAppearance: string;
   status: string;
   category: string;
   universe: string;
}